Extract `useAddToCart`
